Combine logging_handlers.py and logger.py
